refactor: Chip Type 재정의, Item 분리
